+++ devicemanagement/tests/functional/FunctionalTestDevMgmtApiFirmwareMgmtCmds.cpp
@@ -22,14 +22,14 @@ class FunctionalTestDevMgmtApiFirmwareMgmtCmds : public TestDevMgmtApiSyncCmds {
     handle_ = dlopen("libDM.so", RTLD_LAZY);
     devLayer_ = IDeviceLayer::createPcieDeviceLayer(false, true);
     initTestTrace();
-    initEventProcessor();
     controlTraceLogging();
     initDevErrorEvent();
+    // NOTE: DM Event processor cannot run during ETSOC reset
+    // TODO: SW-18858: Add support to stop and restart event processor for ETSOC
+    // reset keeping the option to run the event processor threads in detach mode
   }
   void TearDown() override {
-    cleanupEventProcessor();
-    // NOTE: Skip checking of device error events in ETSOC reset tests because error counters
-    // are also reset during the reset
+    checkDevErrorEvent();
     extractAndPrintTraceData(false /* multiple devices */, TraceBufferType::TraceBufferSP);
     if (handle_ != nullptr) {
       dlclose(handle_);
@@ -38,15 +38,18 @@ class FunctionalTestDevMgmtApiFirmwareMgmtCmds : public TestDevMgmtApiSyncCmds {
 };
 
 TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, getMMErrorCount) {
+  initEventProcessor();
   getMMErrorCount(false /* Multiple devices */);
+  cleanupEventProcessor();
 }
 
 TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, getFWBootstatus) {
   // TODO: SW-13807: Enable back on silicon after fix
   // if (targetInList({Target::FullBoot, Target::Silicon})) {
   if (targetInList({Target::FullBoot})) {
+    initEventProcessor();
     getFWBootstatus(false /* Multiple devices */);
-    checkDevErrorEvent();
+    cleanupEventProcessor();
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -55,8 +58,9 @@ TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, getFWBootstatus) {
 
 TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, getModuleFWRevision) {
   if (targetInList({Target::FullBoot, Target::Silicon})) {
+    initEventProcessor();
     getModuleFWRevision(false /* Multiple devices */);
-    checkDevErrorEvent();
+    cleanupEventProcessor();
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -74,7 +78,9 @@ TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, updateFirmwareImage) {
       FLAGS_enable_trace_dump = false;
       return;
     }
+    initEventProcessor();
     setFirmwareUpdateImage(false /* Multiple Devices */, false);
+    cleanupEventProcessor();
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;
@@ -87,6 +93,9 @@ TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, resetSOCSingleDevice) {
     FLAGS_enable_trace_dump = false;
     return;
   }
+  // NOTE: Skip checking of device error events in ETSOC reset tests because error counters
+  // are also reset during the reset
+  setDevErrorEventCheckList({});
   resetSOC(true /* Single Device */);
 }
 
@@ -96,13 +105,17 @@ TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, resetSOCMultiDevice) {
     FLAGS_enable_trace_dump = false;
     return;
   }
+  // NOTE: Skip checking of device error events in ETSOC reset tests because error counters
+  // are also reset during the reset
+  setDevErrorEventCheckList({});
   resetSOC(false /* Multiple Devices */);
 }
 
 TEST_F(FunctionalTestDevMgmtApiFirmwareMgmtCmds, testShireCacheConfig) {
   if (targetInList({Target::FullBoot, Target::Silicon})) {
+    initEventProcessor();
     testShireCacheConfig(false /* Multiple Devices */);
-    checkDevErrorEvent();
+    cleanupEventProcessor();
   } else {
     DV_LOG(INFO) << "Skipping the test since its not supported on current target";
     FLAGS_enable_trace_dump = false;